FEAT: Added Pending debts in the dashboard and disabled debit if TotalAmount is 0
