+++ Components/Pages/AddTransaction.razor
@@ -4,7 +4,7 @@
     <nav class="sidebar-nav">
         <h1 class="sidebar-title">BudgetMate</h1>
         <ul>
-            <li><a href="/dashboard" > Dashboard</a></li>
+            <li><a href="/dashboard"> Dashboard</a></li>
             <li><a href="/transactions">All Transactions</a></li>
             <li><a href="/transcationForm" class="active"> Add Transaction</a></li>
             <li><a href="/tags">Manage Tags</a></li>
@@ -13,7 +13,6 @@
     </nav>
     <a href="/" class="logout-btn">Logout</a>
 </div>
-
 
 <main class="transaction-form-container">
     <h2>Add Transaction</h2>
@@ -30,10 +29,9 @@
             </div>
             <div class="form-group transaction-type">
                 <label>Transaction Type</label>
-                <select @bind="transaction.Type" >
-                    <option value="----" disabled>Select transaction type</option>
+                <select @bind="transaction.Type">
                     <option value="Credit">Credit</option>
-                    <option value="Debit">Debit</option>
+                    <option value="Debit" disabled="@(TotalAmount <= 0)">Debit</option>
                 </select>
             </div>
             <div class="form-group">
