+++ Components/Pages/AllTansactions.razor
@@ -36,7 +36,7 @@
                 <option value="Debit">Debit</option>
             </select>
             <button class="sort-btn" @onclick="FilterTransactions">Apply Filters</button>
-            <button class="export-btn" @onclick="ExportToJson">Export to JSON</button>
+            <button class="sort-btn" @onclick="ExportToJson">Export to JSON</button>
         </div>
     </header>
