+++ Components/Pages/AllTansactions.razor.css
