+++ Components/Pages/Dashboard.razor.css
@@ -4,7 +4,7 @@
   height: 100vh;
   display: flex;
   flex-direction: column;
-  justify-content: space-between; 
+  justify-content: space-between;
   position: fixed;
   left: 0;
   top: 0;
@@ -14,7 +14,7 @@
 
 .sidebar-title {
   margin-bottom: 1.5rem;
-  font-size: 1.25rem; 
+  font-size: 1.25rem;
   text-align: center;
 }
 
@@ -92,7 +92,7 @@
 
 .stats-grid {
   display: grid;
-  grid-template-columns: repeat(4, 1fr);
+  grid-template-columns: repeat(5, 1fr);
   gap: 1.5rem;
   margin-bottom: 2.5rem;
 }
