+++ wwwroot/Data/debts.json
@@ -30,5 +30,13 @@
     "Lender": "goku",
     "Status": "Pending",
     "Note": "test"
+  },
+  {
+    "Id": 5,
+    "Amount": 0,
+    "Date": "2025-01-17",
+    "Lender": "",
+    "Status": "Pending",
+    "Note": null
   }
 ]
